move validator property fn and tests from pool.go to validator.go
