+++ x/stake/pool.go
@@ -128,118 +128,3 @@ func (p Pool) removeSharesUnbonded(shares sdk.Rat) (p2 Pool, removedTokens int64
 	p.UnbondedTokens -= removedTokens
 	return p, removedTokens
 }
-
-//_______________________________________________________________________
-// Validator Properties TODO move back to types.go under validator (maybe split types.go)
-
-// XXX write test
-// update the location of the shares within a validator if its bond status has changed
-func (p Pool) UpdateSharesLocation(validator Validator) (Pool, Validator) {
-	var tokens int64
-
-	switch {
-	case !validator.BondedShares.IsZero():
-		if validator.Status == sdk.Bonded { // return if nothing needs switching
-			return p, validator
-		}
-		p, tokens = p.removeSharesBonded(validator.BondedShares)
-	case !validator.UnbondingShares.IsZero():
-		if validator.Status == sdk.Unbonding {
-			return p, validator
-		}
-		p, tokens = p.removeSharesUnbonding(validator.BondedShares)
-	case !validator.UnbondedShares.IsZero():
-		if validator.Status == sdk.Unbonding {
-			return p, validator
-		}
-		p, tokens = p.removeSharesUnbonded(validator.BondedShares)
-	}
-
-	switch validator.Status {
-	case sdk.Bonded:
-		p, validator.BondedShares = p.addTokensBonded(tokens)
-	case sdk.Unbonding:
-		p, validator.UnbondingShares = p.addTokensUnbonding(tokens)
-	case sdk.Unbonded, sdk.Revoked:
-		p, validator.UnbondedShares = p.addTokensUnbonded(tokens)
-	}
-
-	return p, validator
-}
-
-// XXX TEST
-// get the power or potential power for a validator
-// if bonded, the power is the BondedShares
-// if not bonded, the power is the amount of bonded shares which the
-//    the validator would have it was bonded
-func (p Pool) EquivalentBondedShares(validator Validator) (power sdk.Rat) {
-	switch validator.Status {
-	case sdk.Bonded:
-		power = validator.BondedShares
-	case sdk.Unbonding:
-		shares := validator.UnbondingShares                           // ubShr
-		exRate := p.unbondingShareExRate().Quo(p.bondedShareExRate()) // (tok/ubshr)/(tok/bshr) = bshr/ubshr
-		power = shares.Mul(exRate)                                    // ubshr*bshr/ubshr = bshr
-	case sdk.Unbonded, sdk.Revoked:
-		shares := validator.UnbondedShares                           // ubShr
-		exRate := p.unbondedShareExRate().Quo(p.bondedShareExRate()) // (tok/ubshr)/(tok/bshr) = bshr/ubshr
-		power = shares.Mul(exRate)                                   // ubshr*bshr/ubshr = bshr
-	}
-	return
-}
-
-// get the equivalent amount of tokens contained by a validator
-func (p Pool) validatorTokens(v Validator) sdk.Rat {
-	switch v.Status {
-	case sdk.Bonded:
-		return p.unbondedShareExRate().Mul(v.BondedShares) // (tokens/shares) * shares
-	case sdk.Unbonding:
-		return p.unbondedShareExRate().Mul(v.UnbondingShares)
-	case sdk.Unbonded, sdk.Revoked:
-		return p.unbondedShareExRate().Mul(v.UnbondedShares)
-	}
-	return sdk.ZeroRat()
-}
-
-// XXX Audit this function further to make sure it's correct
-// add tokens to a validator
-func (p Pool) validatorAddTokens(validator Validator,
-	amount int64) (p2 Pool, validator2 Validator, issuedDelegatorShares sdk.Rat) {
-
-	var poolShares sdk.Rat
-	switch validator.Status {
-	case sdk.Bonded:
-		p, poolShares = p.addTokensBonded(amount)
-		validator.BondedShares = validator.BondedShares.Add(poolShares)
-	case sdk.Unbonding:
-		p, poolShares = p.addTokensUnbonding(amount)
-		validator.UnbondingShares = validator.UnbondingShares.Add(poolShares)
-	case sdk.Unbonded, sdk.Revoked:
-		p, poolShares = p.addTokensUnbonded(amount)
-		validator.UnbondedShares = validator.UnbondedShares.Add(poolShares)
-	}
-
-	equivalentBondedShares := p.EquivalentBondedShares(validator)
-	exRate := validator.DelegatorShareExRate(p) // eq-val-bonded-shares/delegator-shares
-	issuedDelegatorShares = equivalentBondedShares.Quo(exRate)
-	validator.DelegatorShares = validator.DelegatorShares.Add(issuedDelegatorShares)
-
-	return p, validator, issuedDelegatorShares
-}
-
-// remove shares from a validator
-func (p Pool) validatorRemoveShares(validator Validator,
-	shares sdk.Rat) (p2 Pool, validator2 Validator, createdCoins int64) {
-
-	//exRate := validator.DelegatorShareExRate() //XXX make sure not used
-
-	globalPoolSharesToRemove := validator.DelegatorShareExRate(p).Mul(shares)
-	if validator.Status == sdk.Bonded {
-		p, createdCoins = p.removeSharesBonded(globalPoolSharesToRemove)
-	} else {
-		p, createdCoins = p.removeSharesUnbonded(globalPoolSharesToRemove)
-	}
-	validator.BondedShares = validator.BondedShares.Sub(globalPoolSharesToRemove)
-	validator.DelegatorShares = validator.DelegatorShares.Sub(shares)
-	return p, validator, createdCoins
-}
